[minor] Address test compilation warnings

+++ test/common/test_storage_fuzz.cpp
@@ -145,7 +145,7 @@ public:
 		auto it = unflushed_chunks.find(location);
 		if (it != unflushed_chunks.end()) {
 			// Check that if there is an existing chunk - it's size matches exactly
-			REQUIRE(it->second.size() == nr_bytes);
+			REQUIRE(it->second.size() == static_cast<size_t>(nr_bytes));
 			it->second = std::string((char *)buffer, nr_bytes);
 		} else {
 			unflushed_chunks.emplace(location, std::string((char *)buffer, nr_bytes));
@@ -174,7 +174,7 @@ public:
 				PRINT_VERBOSE("FS read cached chunk at offset=" << location << " bytes=" << nr_bytes);
 				const auto &data = it->second;
 				// Assume block-aligned reads
-				REQUIRE(data.size() == nr_bytes);
+				REQUIRE(data.size() == static_cast<size_t>(nr_bytes));
 				memcpy(buffer, data.data(), nr_bytes);
 				return;
 			}
